split topics into separate route, home for regular categories. other misc fixes. breadcrumbs
